Splited controller and services
