+++ EPlast/EPlast.BLL/Services/Auth/AuthService.cs
@@ -16,10 +16,6 @@ using EPlast.BLL.Models;
 using EPlast.DataAccess.Repositories;
 using Newtonsoft.Json;
 using NLog.Extensions.Logging;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Routing;
-using Microsoft.AspNetCore.Mvc.Infrastructure;
 
 namespace EPlast.BLL.Services
 {
@@ -30,29 +26,22 @@ namespace EPlast.BLL.Services
         private readonly IEmailConfirmation _emailConfirmation;
         private readonly IMapper _mapper;
         private readonly IRepositoryWrapper _repoWrapper;
-        private readonly IUrlHelper _Url;
-        private readonly IHttpContextAccessor _contextAccessor;
         public AuthService(
             UserManager<User> userManager,
             SignInManager<User> signInManager,
             IEmailConfirmation emailConfirmation,
             IMapper mapper,
-            IRepositoryWrapper repoWrapper,
-            IUrlHelperFactory urlHelperFactory,
-            IActionContextAccessor actionContextAccessor,
-            IHttpContextAccessor contextAccessor)
+            IRepositoryWrapper repoWrapper)
         {
             _userManager = userManager;
             _signInManager = signInManager;
             _emailConfirmation = emailConfirmation;
             _mapper = mapper;
             _repoWrapper = repoWrapper;
-            _Url = urlHelperFactory.GetUrlHelper(actionContextAccessor.ActionContext);
-            _contextAccessor = contextAccessor;
         }
 
         ///<inheritdoc/>
-        public async Task<Microsoft.AspNetCore.Identity.SignInResult> SignInAsync(LoginDto loginDto)
+        public async Task<SignInResult> SignInAsync(LoginDto loginDto)
         {
             var user = _userManager.FindByEmailAsync(loginDto.Email);
             var result = await _signInManager.PasswordSignInAsync(user.Result, loginDto.Password, loginDto.RememberMe, true);
@@ -122,9 +111,9 @@ namespace EPlast.BLL.Services
         }
 
         ///<inheritdoc/>
-        public async Task<Microsoft.AspNetCore.Identity.SignInResult> GetSignInResultAsync(ExternalLoginInfo externalLoginInfo)
+        public async Task<SignInResult> GetSignInResultAsync(ExternalLoginInfo externalLoginInfo)
         {
-            Microsoft.AspNetCore.Identity.SignInResult signInResult = await _signInManager.ExternalLoginSignInAsync(externalLoginInfo.LoginProvider,
+            SignInResult signInResult = await _signInManager.ExternalLoginSignInAsync(externalLoginInfo.LoginProvider,
                     externalLoginInfo.ProviderKey, isPersistent: false, bypassTwoFactor: true);
             return signInResult;
         }
@@ -227,43 +216,6 @@ namespace EPlast.BLL.Services
         public UserDTO GetUser(User user)
         {
             return _mapper.Map<User, UserDTO>(user);
-        }
-
-        ///<inheritdoc/>
-        public async Task<bool> SendEmailRegistr(string email)
-        {
-            string token = await this.AddRoleAndTokenAsync(email);
-            var user = await _userManager.FindByEmailAsync(email);
-            string confirmationLink = _Url.Action(
-                        action: "confirmingEmail",
-                        controller: "Auth",
-                        values: new { token, userId = user.Id },
-                        protocol: _contextAccessor.HttpContext.Request.Scheme);
-            user.EmailSendedOnRegister = DateTime.Now;
-            return (await _emailConfirmation.SendEmailAsync(
-                email: email,
-                subject: "Підтвердження реєстрації ",
-                message: $"Підтвердіть реєстрацію, перейшовши за :  <a href='{confirmationLink}'>посиланням</a> ",
-                title: "Адміністрація сайту EPlast"));
-        }
-
-        ///<inheritdoc/>
-        public async Task SendEmailReminder(string citiesUrl, UserDTO userDTO)
-        {
-            var user = await _userManager.FindByEmailAsync(userDTO.Email);
-            user.EmailSendedOnRegister = DateTime.Now;
-            await _emailConfirmation.SendEmailAsync(user.Email, "Вітаємо у системі!",
-                $"Ви успішно активували свій акаунт!\nНе забудьте доєднатись до осередку, перейшовши за :  <a href='{citiesUrl}'>посиланням</a> ", "Адміністрація сайту EPlast");
-        }
-
-        ///<inheritdoc/>
-        public async Task SendEmailReseting(string confirmationLink, ForgotPasswordDto forgotPasswordDto)
-        {
-            var user = await _userManager.FindByEmailAsync(forgotPasswordDto.Email);
-            user.EmailSendedOnForgotPassword = DateTime.Now;
-            await _userManager.UpdateAsync(user);
-            await _emailConfirmation.SendEmailAsync(forgotPasswordDto.Email, "Скидування пароля",
-                $"Для скидування пароля перейдіть за : <a href='{confirmationLink}'>посиланням</a>", "Адміністрація сайту EPlast");
         }
 
         ///<inheritdoc/>
